Moved helper into existing helper structure. Added some documentation.
