--- test/TODO
@@ -1,9 +1,0 @@
-
-- documentation in _helpers.js
-- incorporate _helpers.js into util/helper.js or just move to util/server.js
-- modify/parallelise domain tests
-- inject into domain/model where possible/appropriate
-- parties.test.js is in fact a participants test
-- move validation plugin to plugin.js
-- create all mock requests in a helper, in beforeEach, to simplify and focus test content?
-- see TODO in server.js
