+++ test/util/helper.js
@@ -35,6 +35,48 @@ const validatePayeeFspUri = Mustache.render(Config.SWITCH_ENDPOINT + Enums.EndPo
 const defaultSwitchHeaders = resource => defaultHeaders(Enums.Http.HeaderResources.SWITCH, resource, Enums.Http.HeaderResources.SWITCH)
 const getPayerfspEndpointsUri = Mustache.render(Config.SWITCH_ENDPOINT + Enums.EndPoints.FspEndpointTemplates.PARTICIPANT_ENDPOINTS_GET, { fsp: payerfsp })
 const getPayeefspEndpointsUri = Mustache.render(Config.SWITCH_ENDPOINT + Enums.EndPoints.FspEndpointTemplates.PARTICIPANT_ENDPOINTS_GET, { fsp: payeefsp })
+const { createServer, service } = require('../../src/server')
+const getPort = require('get-port')
+const ErrorHandler = require('@mojaloop/central-services-error-handling')
+const DefaultCentral = { ErrorHandler }
+
+const nullLogger = ['error', 'debug', 'warn', 'silly', 'log', 'trace', 'info']
+  .reduce((pv, method) => Object.assign(pv, { [method]: () => {} }), {})
+
+/* @function startTestAdminServer
+ * @function startTestAPIServer
+ *
+ * @description Convenience utilities for instantiating a server and placing it in the test context
+ *
+ * @param {function} buildAppMocks - a function that builds a set of mocks that are injected into
+ *     handlers. These should mock the dependencies that are assigned to server.app in server.js.
+ *
+ * @returns {function} A function that accepts an ava test context, upon which it will store the
+ *     instantiated server.
+ */
+const startTestAdminServer = (buildAppMocks = () => ({})) => startTestServer(service.ADMIN, buildAppMocks)
+const startTestAPIServer = (buildAppMocks = () => ({})) => startTestServer(service.API, buildAppMocks)
+
+const startTestServer = (service, buildAppMocks) => async t => {
+  const appMockDefaults = { logger: nullLogger, Central: DefaultCentral }
+  const appMocks = { ...appMockDefaults, ...buildAppMocks() }
+  t.context.server = await createServer(await getPort(), service, appMocks)
+  // The following code is looking for the Blipp plugin. It looks for the text of the listener
+  // function that Blipp registers against the start event. If the text of that function matches
+  // the regex here, it replaces the listener function to do nothing so it's not a nuisance during
+  // the tests.
+  // This was deemed to be easier than configuring the server create code to allow plugin
+  // configuration- especially when the only intention is for testing.
+  // Are your tests failing because of this code? It was deliberately written to fail
+  // obviously if the conditions leading to its existence were no longer true.
+  // - if you're not using Blipp any more, delete the following statement
+  // - if the content of the Blipp .register function has changed, you may need to change the regex
+  //   here, or delete or comment this line and put up with getting Blipped in your test output
+  const listenerRegex = /const out = .*\n.*console\.log\(out\)/
+  t.context.server.events._eventListeners.start.handlers.find(h => listenerRegex.test(h.listener))
+    .listener = () => {}
+  await t.context.server.start()
+}
 
 /**
  * @function defaultHeaders
@@ -249,5 +291,8 @@ module.exports = {
   getEndPointsResponse,
   fspIdPayload,
   participantPutEndpointOptions,
-  postByTypeIdCurrencyRequest
+  postByTypeIdCurrencyRequest,
+  startTestAPIServer,
+  startTestAdminServer,
+  startTestServer
 }
